+++ src/Git_Updater_PRO/Remote_Management.php
@@ -133,9 +133,13 @@ class Remote_Management {
 		if ( empty( self::$api_key ) ) {
 			$this->load_options();
 		}
-		$api_url = add_query_arg(
+		$update_endpoint       = add_query_arg(
 			[ 'key' => self::$api_key ],
 			home_url( 'wp-json/' . $this->get_class_vars( 'REST\REST_API', 'namespace' ) . '/update/' )
+		);
+		$branch_reset_endpoint = add_query_arg(
+			[ 'key' => self::$api_key ],
+			home_url( 'wp-json/' . $this->get_class_vars( 'REST\REST_API', 'namespace' ) . '/branch-reset/' )
 		);
 
 		echo '<p>';
@@ -148,7 +152,7 @@ class Remote_Management {
 				/* translators: %s: Link to Git Remote Updater repository */
 				__( 'The <a href="%s">Git Remote Updater</a> plugin was specifically created to make the remote management of Git Updater supported plugins and themes much simpler. You will need the Site URL and REST API key to use with Git Remote Updater settings.', 'git-updater-pro' )
 			),
-			'https://github.com/afragen/git-remote-updater'
+			'https://git-updater.com/knowledge-base/git-remote-updater/'
 		);
 		echo '</p>';
 
@@ -167,10 +171,29 @@ class Remote_Management {
 		printf(
 			wp_kses_post(
 				/* translators: 1: Link to wiki, 2: RESTful API URL */
-				__( 'Please refer to the <a href="%1$s">wiki</a> for complete list of attributes. REST API endpoints for webhook updating begin at: %2$s', 'git-updater-pro' )
+				__( 'Please refer to the <a href="%s">Git Updater Knowledge Base</a> for complete list of attributes.', 'git-updater-pro' )
 			),
-			'https://github.com/afragen/git-updater/wiki/Remote-Management---RESTful-Endpoints',
-			'<br><span style="font-family:monospace;">' . esc_url( $api_url ) . '</span>'
+			'https://git-updater.com/knowledge-base/remote-management-restful-endpoints/',
+		);
+		echo '</p>';
+
+		echo '<p>';
+		printf(
+			wp_kses_post(
+				/* translators: link to REST API endpoint for updating */
+				__( 'REST API endpoints for webhook updating begin at: %s', 'git-updater-pro' )
+			),
+			'<br><span style="font-family:monospace;">' . esc_url( $update_endpoint ) . '</span>'
+		);
+		echo '</p>';
+
+		echo '<p>';
+		printf(
+			wp_kses_post(
+				/* translators: link to REST API endpoint for branch resetting */
+				__( 'REST API endpoints for webhook branch resetting begin at: %s', 'git-updater-pro' )
+			),
+			'<br><span style="font-family:monospace;">' . esc_url( $branch_reset_endpoint ) . '</span>'
 		);
 		echo '</p>';
 	}